Restructure `syntax::parse` and remove `ParseError`
We no longer issue parser errors. Get rid of the `error` module from
the parser entirely. Move the parser to the root of the `parse` module
and generally tidy things up a bit.
